+++ src/config/tyractorsaur_config.rs
@@ -1,5 +1,6 @@
 use crate::config::global_config::GlobalConfig;
 use crate::config::pool_config::PoolConfig;
+use crate::config::remoting::RemotingConfig;
 
 use config::{Config, ConfigError, Environment, File, FileFormat};
 use serde::{Deserialize, Serialize};
@@ -11,6 +12,7 @@ pub const SYSTEM_POOL: &str = "system";
 pub struct TyractorsaurConfig {
     pub global: GlobalConfig,
     pub thread_pool: PoolConfig,
+    pub remoting: RemotingConfig
 }
 
 impl TyractorsaurConfig {