+++ src/remoting/mod.rs
@@ -1,0 +1,5 @@
+mod remote_actor;
+
+pub mod prelude {
+    pub use crate::remoting::*;
+}