+++ src/router/mod.rs
@@ -1,0 +1,5 @@
+mod round_robin_router;
+
+pub mod prelude {
+    pub use crate::router::round_robin_router::*;
+}